refactor(api): move command related API to separate file
